{{cookiecutter.project_slug}}/requirements/main.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=- requirements/main.in
#
arrow==1.0.3
    # via jinja2-time
attrs==20.3.0
    # via pytest
bandit==1.7.0
    # via -r requirements/main.in
binaryornot==0.4.4
    # via cookiecutter
blessings==1.7
    # via curtsies
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via
    #   binaryornot
    #   requests
click==7.1.2
    # via cookiecutter
cookiecutter==1.7.2
    # via -r requirements/main.in
coverage==5.5
    # via pytest-cov
curtsies==0.3.5
    # via enquiries
cwcwidth==0.1.4
    # via curtsies
docopt==0.6.2
    # via -r requirements/main.in
enquiries==0.1.0
    # via -r requirements/main.in
gitdb==4.0.5
    # via gitpython
gitpython==3.1.14
    # via bandit
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
jinja2-time==0.2.0
    # via cookiecutter
jinja2==2.11.3
    # via
    #   cookiecutter
    #   jinja2-time
markupsafe==1.1.1
    # via
    #   cookiecutter
    #   jinja2
packaging==20.9
    # via pytest
pbr==5.5.1
    # via stevedore
pluggy==0.13.1
    # via pytest
poyo==0.5.0
    # via cookiecutter
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements/main.in
pytest-html==3.1.1
    # via -r requirements/main.in
pytest-metadata==1.11.0
    # via pytest-html
pytest==6.2.2
    # via
    #   -r requirements/main.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
python-dateutil==2.8.1
    # via arrow
python-slugify==4.0.1
    # via cookiecutter
pyyaml==5.4.1
    # via
    #   -r requirements/main.in
    #   bandit
requests==2.25.1
    # via cookiecutter
six==1.15.0
    # via
    #   bandit
    #   blessings
    #   cookiecutter
    #   python-dateutil
smmap==3.0.5
    # via gitdb
stevedore==3.3.0
    # via bandit
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via pytest
urllib3==1.26.3
    # via requests
